OV-326: * unify references of modal in one place, move logic of close to video-menu
