+++ frontend/src/bundles/studio/components/video-menu/components/menu-body/menu-body.tsx
@@ -6,8 +6,6 @@ import {
     Icon,
     IconButton,
 } from '~/bundles/common/components/components.js';
-import { DOM_EVENT } from '~/bundles/common/enums/enums.js';
-import { useEffect, useRef } from '~/bundles/common/hooks/hooks.js';
 import { IconName } from '~/bundles/common/icons/icons.js';
 
 import styles from './styles.module.css';
@@ -16,8 +14,7 @@ type Properties = {
     title: string | React.ReactNode;
     onClose: () => void;
     onChatOpen?: () => void;
-    chatModalReference: React.RefObject<HTMLDivElement>;
-    scriptModalReference: React.RefObject<HTMLDivElement>;
+    menuBodyReference: React.RefObject<HTMLDivElement>;
 };
 
 const MenuBody: React.FC<React.PropsWithChildren<Properties>> = ({
@@ -25,42 +22,8 @@ const MenuBody: React.FC<React.PropsWithChildren<Properties>> = ({
     children,
     onClose,
     onChatOpen,
-    chatModalReference,
-    scriptModalReference,
+    menuBodyReference,
 }) => {
-    const menuBodyReference = useRef<HTMLDivElement | null>(null);
-
-    useEffect(() => {
-        const handleClickOutside = (event: MouseEvent): void => {
-            const isInsideMenuBody = menuBodyReference.current?.contains(
-                event.target as Node,
-            );
-            const isInsideChatModal = chatModalReference.current?.contains(
-                event.target as Node,
-            );
-            const isInsideScriptModal = scriptModalReference.current?.contains(
-                event.target as Node,
-            );
-
-            if (
-                !isInsideMenuBody &&
-                !isInsideChatModal &&
-                !isInsideScriptModal
-            ) {
-                onClose();
-            }
-        };
-
-        document.addEventListener(DOM_EVENT.MOUSE_DOWN, handleClickOutside);
-
-        return () => {
-            document.removeEventListener(
-                DOM_EVENT.MOUSE_DOWN,
-                handleClickOutside,
-            );
-        };
-    }, [onClose, menuBodyReference, chatModalReference, scriptModalReference]);
-
     return (
         <Box
             ref={menuBodyReference}
