+++ frontend/src/bundles/studio/components/video-menu/video-menu.tsx
@@ -1,5 +1,6 @@
 import { ChatModal } from '~/bundles/chat/pages/chat-modal.js';
 import { Icon } from '~/bundles/common/components/components.js';
+import { DOM_EVENT } from '~/bundles/common/enums/enums.js';
 import {
     useAppDispatch,
     useAppSelector,
@@ -30,8 +31,8 @@ const VideoMenu: React.FC = () => {
 
     const dispatch = useAppDispatch();
     const [isChatOpen, setIsChatOpen] = useState(false);
-    const chatModalReference = useRef<HTMLDivElement | null>(null);
-    const scriptModalReference = useRef<HTMLDivElement | null>(null);
+    const modalReference = useRef<HTMLDivElement | null>(null);
+    const menuBodyReference = useRef<HTMLDivElement | null>(null);
 
     const setActiveItem = useCallback(
         (item: ValueOf<typeof MenuItems> | null): void => {
@@ -77,9 +78,7 @@ const VideoMenu: React.FC = () => {
         script: {
             label: 'Script',
             icon: <Icon as={IconName.SCRIPT} />,
-            getContent: () => (
-                <ScriptContent modalReference={scriptModalReference} />
-            ),
+            getContent: () => <ScriptContent modalReference={modalReference} />,
         },
         // text: {
         //     label: 'Text',
@@ -95,6 +94,30 @@ const VideoMenu: React.FC = () => {
 
     const activeMenuItem = activeItem && menuItems[activeItem];
 
+    useEffect(() => {
+        const handleClickOutside = (event: MouseEvent): void => {
+            const isInsideMenuBody = menuBodyReference.current?.contains(
+                event.target as Node,
+            );
+            const isInsideModal = modalReference.current?.contains(
+                event.target as Node,
+            );
+
+            if (!isInsideMenuBody && !isInsideModal) {
+                handleMenuClose();
+            }
+        };
+
+        document.addEventListener(DOM_EVENT.MOUSE_DOWN, handleClickOutside);
+
+        return () => {
+            document.removeEventListener(
+                DOM_EVENT.MOUSE_DOWN,
+                handleClickOutside,
+            );
+        };
+    }, [handleMenuClose, menuBodyReference, modalReference]);
+
     return (
         <>
             <Menu
@@ -107,8 +130,7 @@ const VideoMenu: React.FC = () => {
                     title={activeMenuItem.label}
                     onClose={handleMenuClose}
                     onChatOpen={handleChatOpen}
-                    chatModalReference={chatModalReference}
-                    scriptModalReference={scriptModalReference}
+                    menuBodyReference={menuBodyReference}
                 >
                     {activeMenuItem.getContent()}
                 </MenuBody>
@@ -116,7 +138,7 @@ const VideoMenu: React.FC = () => {
             <ChatModal
                 isChatOpen={isChatOpen}
                 onModalChatClose={handleChatClose}
-                modalReference={chatModalReference}
+                modalReference={modalReference}
             />
         </>
     );
